refactor(domain): PascalCase files, pure domain (codec→infra), fold takes Seq
レビュー指摘の反映:
- Scala ファイルを型名対応の PascalCase に（ArticleId/Slug/NonEmptyTitle/ArticleEvent/Article.scala）。
  AGENTS.md の命名規約も「Scala は PascalCase、TS/ディレクトリは kebab-case」に明確化
- domain から JSON codec(zio-json/iron-zio-json)を除去し純粋化（serialization は infrastructure の関心事）。
  domain の依存は iron(モデリング) + zio-test のみに。codec はタスク2で infra に実装
- fold の引数を Istable→Seq に（イベント順序前提を型で強制。codex Low 指摘）
- domain テストは VO + fold の 7 件（codec round-trip はタスク2へ）
- README / design.md / slice-1-plan.md を更新
