+++ apps/api/build.sbt
@@ -3,7 +3,6 @@
 
 val scala3      = "3.8.4"
 val zioVersion  = "2.1.26"
-val zioJson     = "0.9.2"
 val ironVersion = "3.3.1"
 
 ThisBuild / scalaVersion := scala3
@@ -17,19 +16,18 @@ ThisBuild / scalacOptions ++= Seq(
   "-Wunused:all",
 )
 
-// 純粋ドメイン: Article イベント ADT・値オブジェクト(Iron)・projection fold・JSON codec。
-// IO(Firestore/Postgres/tapir)への依存を載せないことをビルドで強制する。
-// RMU(Rust) とはコードを共有せず、イベント wire スキーマを契約フィクスチャで整合させる。
+// 純粋ドメイン: Article イベント ADT・値オブジェクト(Iron)・projection fold。
+// IO や wire フォーマット(JSON 等)への依存を載せないことをビルドで強制する。
+// JSON codec は infrastructure 層の関心事。RMU(Rust) とはコードを共有せず、
+// イベント wire スキーマを契約フィクスチャで整合させる。
 lazy val domain = project
   .in(file("domain"))
   .settings(
     name := "morecat-domain",
     libraryDependencies ++= Seq(
-      "dev.zio"            %% "zio-json"      % zioJson,
-      "io.github.iltotore" %% "iron"          % ironVersion,
-      "io.github.iltotore" %% "iron-zio-json" % ironVersion,
-      "dev.zio"            %% "zio-test"      % zioVersion % Test,
-      "dev.zio"            %% "zio-test-sbt"  % zioVersion % Test,
+      "io.github.iltotore" %% "iron"         % ironVersion,
+      "dev.zio"            %% "zio-test"     % zioVersion % Test,
+      "dev.zio"            %% "zio-test-sbt" % zioVersion % Test,
     ),
     testFrameworks += new TestFramework("zio.test.sbt.ZTestFramework"),
   )
