+++ apps/api/domain/src/main/scala/morecat/domain/ArticleId.scala
@@ -1,9 +1,8 @@
 package morecat.domain
 
 import java.util.UUID
-import zio.json.*
 
-/** 集約 ID。アプリ採番の UUID（v7 を想定。生成は infrastructure 層の責務）。 */
+/** 集約 ID。アプリ採番の UUID（v7 を想定。採番・v7 検証は infrastructure 層の責務）。 */
 opaque type ArticleId = UUID
 
 object ArticleId:
@@ -14,6 +13,3 @@ object ArticleId:
     catch case _: IllegalArgumentException => Left(s"invalid UUID: $s")
 
   extension (id: ArticleId) def value: UUID = id
-
-  given JsonCodec[ArticleId] =
-    JsonCodec(summon[JsonEncoder[UUID]], summon[JsonDecoder[UUID]]).transform(apply, _.value)
